requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.75.0
fastapi-jwt-auth==0.5.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.1.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-decouple==3.6
python-jose==3.3.0
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
twilio==6.0.0
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
zipp==3.7.0
